# docs/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "oasis_sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Dashboard and back-office API. Errors are application/problem+json; money is integer cents; instants are ISO-8601 UTC; business dates are YYYY-MM-DD in the business timezone.",
    "title": "Oasis Auto Spa API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/invite/accept": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "type": "string"
                  },
                  "password": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  },
                  "token": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "token",
                  "email",
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "csrfToken": {
                      "type": "string"
                    },
                    "user": {
                      "additionalProperties": false,
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "employeeId": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "employeeId",
                        "email",
                        "name"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "user",
                    "csrfToken"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [],
        "summary": "Accept an invite: set the email and password, activate the employee and sign in",
        "tags": [
          "auth"
        ],
        "x-oasis-access": "public"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "maxLength": 254,
                    "minLength": 3,
                    "type": "string"
                  },
                  "password": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "csrfToken": {
                      "type": "string"
                    },
                    "user": {
                      "additionalProperties": false,
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "employeeId": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "employeeId",
                        "email",
                        "name"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "user",
                    "csrfToken"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [],
        "summary": "Sign in with email and password; sets the session cookie",
        "tags": [
          "auth"
        ],
        "x-oasis-access": "public"
      }
    },
...
    "/api/v1/auth/password/change": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "currentPassword": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  },
                  "newPassword": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "ok": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Change the password; every other session is revoked",
        "tags": [
          "auth"
        ],
        "x-oasis-access": "authenticated"
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "maxLength": 254,
                    "minLength": 3,
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "accepted": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "accepted"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [],
        "summary": "Request a password-reset link (always 202)",
        "tags": [
          "auth"
        ],
        "x-oasis-access": "public"
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "password": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  },
                  "token": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "token",
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "ok": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [],
        "summary": "Set a new password with a reset token; revokes every session of the account",
        "tags": [
          "auth"
        ],
        "x-oasis-access": "public"
      }
    },
...
    "/api/v1/me/preferences": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "theme": {
                    "enum": [
                      "light",
                      "dark"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "theme"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "theme": {
                      "anyOf": [
                        {
                          "enum": [
                            "light",
                            "dark"
                          ],
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "theme"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Save preferences (theme)",
        "tags": [
          "auth"
        ],
        "x-oasis-access": "authenticated"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ]
}